添加快速Cocluster Pipeline使用指南和示例代码
- 新增README.md文件，包含快速开始、基本使用、高级配置、性能优化、结果分析、错误处理、测试建议、最佳实践和常见问题等内容。
- 更新Coclusterer、Config、Pipeline、Scoring等模块，支持使用f64类型，提升数值精度。
- 实现了Pipeline的构建和运行逻辑，支持自定义聚类器和评分器。
- 添加了多种评分器的实现，包括Pearson、Exponential和Compatibility评分器。
- 引入了并行处理和统计信息收集功能，优化了性能。

变更文件：
- src/cocluster.rs
- src/config.rs
- src/lib.rs
- src/matrix.rs
- src/pipeline.rs
- src/scoring.rs
- src/spectral_cocluster.rs
- src/submatrix.rs
- 新增README.md文件

+++ src/config.rs
@@ -5,7 +5,7 @@ use ndarray_npy::ReadNpyExt;
  * Created Date: Friday, January 26th 2024
  * Author: Zihan
  * -----
- * Last Modified: Monday, 29th January 2024 1:40:46 pm
+ * Last Modified: Monday, 26th May 2025 11:31:24 am
  * Modified By: the developer formerly known as Zihan at <wzh4464@gmail.com>
  * -----
  * HISTORY:
@@ -17,7 +17,7 @@ use std::fs::File;
 pub struct Config {
     // 字段定义
     // need a matrix to init, float
-    matrix: Array2<f32>,
+    matrix: Array2<f64>,
     // shape of matrix
     row: usize,
     col: usize,
@@ -25,7 +25,7 @@ pub struct Config {
     m: usize,
     n: usize,
     // torlerance
-    tol: f32,
+    tol: f64,
 }
 
 impl Config {
@@ -47,10 +47,10 @@ impl Config {
         // read from npy file
         args.next();
         let reader = File::open(args.next().unwrap())?;
-        let matrix = Array2::<f32>::read_npy(reader)?;
+        let matrix = Array2::<f64>::read_npy(reader)?;
         let m = args.next().unwrap().parse::<usize>()?;
         let n = args.next().unwrap().parse::<usize>()?;
-        let tol = args.next().unwrap().parse::<f32>()?;
+        let tol = args.next().unwrap().parse::<f64>()?;
         let row = matrix.shape()[0];
         let col = matrix.shape()[1];
 
@@ -64,7 +64,7 @@ impl Config {
         })
     }
 
-    pub fn get_matrix(&self) -> &Array2<f32> {
+    pub fn get_matrix(&self) -> &Array2<f64> {
         &self.matrix
     }
 
@@ -76,7 +76,7 @@ impl Config {
         self.n
     }
 
-    pub fn get_tol(&self) -> f32 {
+    pub fn get_tol(&self) -> f64 {
         self.tol
     }
 
@@ -97,7 +97,7 @@ mod tests {
     fn test_new_config() {
         // make npy file at target/debug/fast_cocluster/test.npy
         let file = File::create("target/debug/test.npy").unwrap();
-        ndarray_npy::WriteNpyExt::write_npy(&ndarray::Array2::<f32>::zeros((24000, 768)), file)
+        ndarray_npy::WriteNpyExt::write_npy(&ndarray::Array2::<f64>::zeros((24000, 768)), file)
             .unwrap();
         let args = vec![
             "target/debug/fast_cocluster".to_string(),
